Removed forward slash
I checked out the code and tried to run localhost and it did not work
for me.  I removed the / in front of assets and it worked for me.  I am
using a Mac computer on El Capitan.  I would like to check on a windows
machine as soon as possible to confirm but we should be good.